Fix new connections with no collapse features (#19)
I forgot about truncation when I introduced the "Not Collapsable" style, and no unit tests picked it up. Any new connection between nodes with an uncollapsable feature would assert.

The truncation algorithm in the EditTree needs to know about expanded nodes, but the style-based expansion was entirely managed by the ContentsCache. In this change, the ContentsCache informs the EditTree when it encounters features which have a style which says they should always be expanded. I don't love the bi-directional dependency, but the change itself is relatively straightforward, so I'll accept it.


* Inform the EditTree of uncollapsable features.

* EditTreeTest truncatePathsWithImplicitlyExpandedPaths