Complex types (WIP) (#23)
Add support for non-primitive types to the type system. (Note that the type system is currently only used by the map system but I expect its scope to expand over time.)

TypeConstructors can construct types from other types. TypeRef replaces Identifier as the way to identify a type, since it can describe a type constructed from other types as well as a primitive type.

The motivation for this change now is to support AddBlank, which automatically build an enum from another enum by adding a blank value. This can be used in the map editor to signal that the source value is to be treated as discarded. (This functionality could have been implemented with a map specific change, but a generic approach was much more interesting to code and will be more valuable in the future.)

In the future, type constructors should be able to support tuple types, map types, etc.

Note: This initial implementation is likely to be inefficient:

    TypeRef is bigger than it should be.
    Lots of small allocations due to vectors.

* Reimplement UUID using uint

* Move hash.hpp into own folder

* Uuid::serializeToString

* Use an unordered_map for Registry

* Small registry simplification

* Stop templating on untyped registry

* Remove onRegistered()

* Big change: Compiling but not linking

* Use a vector for constructed types.

* Transition code to TypeRef

* Now linking.

* Implement some TypeRef methods.

* TypeRef serialization

* Fixes

* equality test

* Fix tests

* Implement operators directly from variant

* visit identifiers test.

* Fix serialization bug

* Hash test

* toString test

* Start of resolve tests

* Allow type constructors to have format strings as names.

* 10 type arguments

* Don't use registry

* Sketch of TypeConstructor

* Move s_maxNumArguments

* Resolve tests for constructed Types.

* First pass at EnumWithDummy.

* Reg AddDummy. Map fix.

* File should be AddDummy

* MapApplicator helper for enum indices

* AddDummy -> AddBlank

* ConstructedType template

* Macro tweaks

* First pass at refactored subtyping: Test failing

* TypeRef::IsSubTypeHelper improvement

* AddBlank subtype fix

* Map serialization fix

* Fix. MapFeature entirely determines type choices.

* SubtypeOrder in its own file. Renames.

* Symmetry. Renames

* TypeSystemTest compareSubtypeComplex

* TypeSystemTest, compareSubtypeBinary

* Move overloaded template to types.hpp

* TypeSystemTest compareSubtypeComplex

* Some Id -> Ref renaming

* More id -> ref renaming

* Serialize with [ and ].

* Comments

* Make constructType protected

* AddBlank tests

* AddBlank -> AddBlankToEnum

* Restrict identifiers to normal rules.

* Use ' to mark the discriminator.

* Hopefully fix CI