Improved TypeRef (#24)
Two changes:

    Identifiers in the type system are now Medium length (two words)
    Replace vector-based storage in TypeRef by variant branches holding strictly-sized arrays of arguments.

The new storage implementation means:

    TypeRef is smaller and tidier.
    Very basic testing suggests a noticeably performance improvement.
    The caches carried by TypeConstructors no longer have vector keys (which was making me sad)

A mild disadvantage is that a TypeRef cannot no longer describe a constructed type with more than 2 arguments. I've no use case for that, but it's nice to be able to represent data that could in theory be valid even if it currently isn't. This now gives a parse error, where previously it would just have some data which represents a unresolving typeref. I considered and discounted having fallback cases for this.

* Identifier -> ShortId

* LongIdentifier -> LongId

* Rename registry methods

* REGISTERED_ID renames

* Start adding support for MediumIds

* Use MediumId for Type Identifiers

* Transition to arrays from vectors

* Templated typeRef

* Reimplement TypeRef visitors

* Tweaks

* Merge fix