Distinguish between regular Values and EditableValues (#26)
Values that can be directly edited within BabelWires need to be serializable, since they can be carried by modifiers. However, BabelWires should be able to handle values that are not directly editable and can only be loaded, processed and saved. An example of the latter is Track from SeqWires.

This PR makes a static distinction between Values and EditableValues, which adds a lot of complexity to ValueHolder.

It also adds UI changes needed to ensure that ValueModels are fully featured.

* Distinguish between EditableValues and non-EditableValues.

* Fix serialization parent

* Move visit functions to EditableValue

* Fix rational serialization

* Introduce ValueModelRegistry

* Add valueModelWithRichText

* Better handle style in ValueModels.

* Add tooltip support for values.