Type based subtyping (#30)
I was attempting to do subtyping algebraically, i.e based on rules involving type constructors. That's not a good fit for the way this type system is shaping up because I allow multiple type constructors to create instances of the same type class.

Instead, let the Type objects do the subtype test.

Also, allow types of carry tags, as an extensible way of grouping them into categories. Subtyping (in the type system OR c++) doesn't seem to be a perfect fit for this.

* Stop trying to do algebraic subtyping

* Add Type::Tag

* Remove related type functionality

* Tag test

* Comment