Instance system (convenience classes for typed traversal of ValueFeature hierarchies) (#35)
Instance system (convenience classes for typed traversal of ValueFeature hierarchies

This change carries Improvements to the type system to support type usage in client code (e.g. the SMF plugin in SeqWires).

In particular, this change adds the Instance system. These provide convenience classes which allow typed traversal of feature hierarchies of value type. The core handling of a project doesn't need these, because it works generically. However, client code (e.g. importer code in a plugin or processor code) knows the structure of the data, so it would be frustrating to explore it generically.


* Add "isNew" to simpleValueFeature as a temporary hack while I consider options.

* Better comment

* RecordFeatureUtils

* Accessor macros for record fields

* missing include

* Add a sensible assert

* More accessors. Add pragma once.

* Return const reference

* arrayFeatureUtils

* FeatureWrappers (requires C++20)

* FeatureWrappers handle optionals.

* arrayFeatureWrapper

* Macro for arrayFeatureWrappers

* RecordWithVariants wrappers

* Fix RecordWithVariantsValue

* Drop operation bool

* Remove old FIELD macros

* Collect wrapper files into InstanceOf folder

* Enum instance of

* FeatureWrapper -> Instance

* InstanceOf -> Instance

* FeatureWrapper->Instance

* No special macro required for recordWithVariants

* Start tidying up instance code

* Further tidies

* Array uses Common base class

* Documentation

* Instance takes a reference

* MSCV fix