Edits at root (#41)
The root of FeatureElements was usually hidden, so it could not be edited and there was therefore no need to store modifiers at the empty path. Additionally, there was no default constructable PathStep to assign to that node.

I now intend to place value types at the root and I want them to be editable.

This PR prepares the way by adding an extra root node to the EditTree to allow it to store modifiers at the empty path. It also introduces a "NotAStep" value for PathStep which is used for the root.

* PathStep now supports a "notAStep" value.

* Make it easier to observe NotAStep

* WIP: EditTree with root edits

* Move RootedPath method bodies into the cpp.

* Ensure root nodes are expanded.

* Validate the root node

* Test expand at root.