WIP: FeatureElements with roots (#42)
I want to delegate the feature system to just being wrappers for values. This PR allows ValueElements to hold values directly, without needing a root feature.

Additionally: The UI now shows the base feature of each element. This is necessary to allow the these roots to be wired to each other.

The UI of the root will need work after this change, but I'm prioritizing the goal of switching from features to values.

* Fix test

* Add root and fix test

* Remove m_indexOffset

* Add a FailedType (WIP)

* ValueElement now has a value as a root

* Remove some unused declarations and includes.

* Give roots a nicer label.

* Use "setInternalFailure" for types which do not resolve

* Don't use "inlined" for roots.

* Expand the root by default