Drop FileFeature (the last non-ValueFeature) (#44)
FileFeature was the last remaining special-purpose feature. This PR removes that in favour of FileType.

I have not attempted to make a value correspond to a File: it makes more sense for files to be managed at the element level, not the value level. Instead, FileType is really just a dummy which reserves a place in the feature hierarchy.

The RowModel abstraction is now almost redundant: It is just used to special case file handling. There's probably a better way of doing this, but for now I've kept it and just made the minimum change so it uses values.

* Remove unused class

* Remove some unused includes

* Temporarily add SourceFileFormat2 and use it for SMF

* Remove some unused includes

* Add FileTypeConstructor

* Switch SourceFileFormat to values.

* Switch TargetFileFormat to using values

* Dispatch to FileRowModel based on type

* Fix bugs due to second registry and dispatching of failure

* Remove unused predeclarations

* Transition tests away from TestFileFeature

* Fix remaining tests

* Remove FileFeature