Support Variadic Types in the TypeConstructor NAME syntax (#55)
Sum and Tuple types are variadic and the syntax for type constructor names meant that the TypeRef::toString could not build a proper string for them.

This PR provides a richer syntax that allows for variadic types and also makes a cleaner distinction between type and value arguments.

See the comments of TypeNameFormatter for details.

* Richer TypeConstructor NAME syntax

* No need to check argument limit here.