Limit the vertical size of the DropDownValueEditor popup (#56)
There was no vertical limit to the size of the popup produced by QComboBox if it occurred within the QGraphicsView. It was OK if it occurred within the map editor. This PR ensures that there's a maximum size and scroll bars are used to move within it.

The new handling works OK, but doesn't look perfect. At least it is now possible to select all pitches in SeqWires' SplitAtPitchProcessor.

As commented within the code, there seems to be a (Qt?) bug where the move doesn't quite place the popup in the correct vertical position, despite the horizontal position being correct. This is particularly visible when zoomed out.

* Warning fix

* Limit the size of QComboBoxes, even when in the graphics view.

* Ensure the combobox popup gets a scroll bar when necessary

* Clearer code. Note about possible Qt bug.