Some identifier improvements (#61)
Unresolved identifiers were designed to resolve automatically after deserialization. This would allow files to be loaded without discriminators, which I thought might be good for external scripting. Unfortunately, the engineering cost of the approach I was using was too high. This PR abandons the use of mutable discriminators. Discriminator handling is more explicit now and the code is better.

As of this PR, unresolved identifiers can be used in a file, but they do not get resolved. (I.e. they will still be unresolved when the data is resaved.) I intend to restore the old behaviour eventually, but I'll implement it in a less magical way.

This PR also includes some other improvements connected to identifiers:
* The Path class is no longer intimately linked to ValueTreeNodes. Traversal methods have moved to utility functions.
* The IdentifierVisitor has been broken out and it's clearer that you don't have to derive from DataVistable to use it.

* Discriminator is no longer mutable.

* PathStep constructor no longer explicit. Remove redundant constructor calls.

* Remove unused method

* Register the identifiers in TestEnum

* Give RecordType an explicitly subtyping constructor.

* Identifier equality is exact on resolved identifiers.

* Introduce ValueTreePathUtils

* Move tryFollow out of path

* Move follow out of Path

* Move ValueTreeNode constructor out of Path

* Fix != and IdentifierTest

* tryFollow -> tryFollowPath

* Path gets  a visitIdentifiers method

* Actually, it's the visitors that should be in their own header.

* More detailed comment on the IdentifierVisitor.

* Use a concept to police IdentifierVisitor usage.

* Update .gitignore with .vscode

* Remove old copyDiscriminator methods

* Stop attempting to explicitly check discriminators in TestProjectData

* Fix RecordType subtype constructor and test.