File Formats should take a path (#63)
It's much more flexible if file formats take a file path rather than an abstraction defined by this codebase. If a format happened to wrap an external API, there might be no common representation of file contents. Adapters could probably be written but using a path seems general and simple, even if it does lock the semantics firmly to the filesystem.

* FileFormats should use a path.

* TargetFileFormat should use path

* Neaten up the test formats

* Better error handling

* Latest SeqWires