Connectability using subtyping (#66)
In nodeeditor v2, connections were permitted based on datatype equality, which basically meant equality of strings. This was very rigid and meant, for example, that summand types could not connect to their sumtype.

With nodeeditor v3 came the ability to programmatically decide whether connections are allowed. Naturally, we use the subtype relation for this.

Unlike many programming systems, we allow connections when types intersect rather than require strict subtyping between source and target. This is because the user is dealing with actual data and can therefore make a judgement about whether to make a connection even if that connection might later be invalidated. To achieve this, the SubtypeOrder is enhanced with a new "IsIntersecting" value.

Note: A negative side-effect of the use of intersection is that false positive connections are possible.

Although IsIntersecting becomes the important relationship in the UI, the compareSubtype implementations still try to return strong values ("subtype", "supertype", "equivalent") if they can. That is needed for certain tasks within the system. In future, an optional "strict mode" might require subtyping, or perhaps future UI support could highlight contingently-valid connections.

* Remove an old TODO

* Make the subtypeOrder more flexible.

* Remove AddBlankToEnum

* Fixes and test fixes

* Use the typeSystem to allow or disallow connections

* Use standard babelwires naming conventions in NodeNodeModel.

* Sum type fix

* subtypeFromRanges fix

* Possibly plausible compareSubtypeHelper algorithm for sumtypes.

* Better sumtype unit test

* Renames and tidies

* Sumtype associativity

* subtypeProduct tests

* Add subtypeUtils

* sumTypes compareSubtypeHelper needs to be tentative if the result is disjoint

* Fix connecting sumtypes

* Update TODOs

* Comment tweaks