RecordTypeConstructor - construct record types at runtime (#37)
Adds a type constructor so RecordTypes can be created in code.

Note: RecordTypes expect fieldIds to be registered, so this PR does not enable record types to be purely created at runtime. Truly dynamic records are definitely interesting, but the implications for data robustness would need to be considered. (As an example of a dynamic type: It might be possible to interpret a JSON schema as a BabelWires type, and then load/modify/save instances of that schema as types within BabelWires. In this particular case, it may even be possible to generate IDs (including UUIDs) for the identifiers if the schema is stable and has a unique "$id" field).

* Add RecordTypeConstructor

* Fix missing include after rebase

* Add IdentifierValueBase and FieldIdValue

* Add makeTypeRef and tests

* Identifiers must be registered