Rename PRIMITIVE_TYPE to REGISTERED_TYPE (#73)
I was misusing the word "Primitive" for types which were registered in the type system. If I wanted to use the term "primitive" then the term should means something like "non-compound" or perhaps refer specifically to non-compound built-in types.

I was using it for types that are registered directly in the type system vs constructed from type constructors, so I've switched to using the more appropriate term "Registered Type".

* PrimitiveType -> RegisteredType

* More renames

* Reference main branch

* Remove unintended change