Connections to compounds with unassigned type variables (#82)
When I introduced generic types, I prevented connections in the project
being made to compound values containing unassigned type variables. This
was purely to limit the implementation complexity. You could connect
directly to a type variable since it was easy to determine the type that
should instantiate the variable.

This PR adds the algorithm needed to determine how type variables should
be instantiated when a compound connection is made, so the restriction
is lifted.

An important change is that the (previously virtual) "isValidValue"
method of types is now implemented in terms of a more flexible virtual
"visitValue" method. There was previous no way to explore a value that
may share some of, but not all of, the appropriate structure for its
type. This functionality is now provided by visitValue and is made use
of by the improved type variable instantiation algorithm. (Most likely,
there are other possible generalizations and simplifications across the
set of Type's virtual functions, but that's outside the scope of this
PR.)